requirements.txt
alembic==1.4.1
Babel==2.8.0
blinker==1.4
cffi==1.14.0
Click==7.0
dnspython==1.16.0
dominate==2.5.1
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.8.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-MQTT==1.0.5
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gevent==1.4.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
monotonic==1.5
mysqlclient==1.4.6
paho-mqtt==1.5.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.0
pytz==2019.3
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.13
typing==3.7.4.1
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
